Revert "Change bot status, added poncho command and grouped commands that post a random pic or message"
This reverts commit a5cafac17043d2af1f7ae26c1ae8ed8056f836da.
